Transform node to configured projection in Deleter
The nodes were passed along to expire in the projection they are in OSM
since they are directly looked up from the cache.
However this results in expire being called with different projections
in deleter and writer since writer correctly reprojects to the
configured projection. Therefore Deleter needs to reproject as well.