postgis: create one composite PRIMARY KEY instead of two indices
Create a composite PRIMARY KEY for OSM ID and the unique `id` colum
(serial) instead of a separeate index for `id` and OSM ID. This can save
a few GB for large tables.

OSM ID is first, so lookups for diff imports are still fast. The `id`
makes the index unique as OSM IDs are not unique (e.g. multiple
mappings).